+++ kauldron/kd.py
@@ -33,6 +33,7 @@ del pytest, sys
 from kauldron import checkpoints as ckpts
 from kauldron import data
 from kauldron import evals
+from kauldron import klinen as knn
 from kauldron import konfig
 from kauldron import kontext
 from kauldron import losses