+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog follow the https://keepachangelog.com/ standard (at least the headers)
 
 ## [Unreleased]
 
+* Add `kd.knn.convert` to convert any Flax module to klinen.
 * Add `kontext.path_builder_from` to dynamically generate keys for the config
   with auto-complete and static type checking.
 * Add `kd.data.BatchSize(XX)` util